The method of obtaining an abstract is combined with the main method 'get_work'. In the files CSV and JSON, the date of file creation is added to the name. The file CSV has a limit of 100,000 rows. The maximum number of works obtained is set to 50 pieces.
