+++ src/server.py
@@ -14,6 +14,7 @@ from src.utils.greeting import hellow_names
 from src.utils.save_fun import save_to_json, append_to_csv
 from src.utils.api_calls import fetch_json
 from config import JSON_SAVE_PATH, CSV_SAVE_PATH, OPENALEX_API_URL
+from config import save_to_json, append_to_csv, fetch_json, remove_latex_expressions, get_abstract_text
 
 if public_or_local == 'LOCAL':
     url = 'http://localhost'
@@ -81,91 +82,33 @@ async def inputation(body: Annotated[
 class WorkRequest(BaseModel):
     work_ids: list[str]
 
-    class Config:
-        json_schema_extra = {
-            "example": {
-                "work_ids": ["W2117692326", "W1234567890", "W0987654321"]
-            }
-        }
-
-def save_to_json(data, work_id):
-    if not os.path.exists(JSON_SAVE_PATH):
-        os.makedirs(JSON_SAVE_PATH)
-    json_filename = os.path.join(JSON_SAVE_PATH, f"{work_id}.json")
-    if not os.path.exists(json_filename):
-        with open(json_filename, 'w', encoding='utf-8') as json_file:
-            json.dump(data, json_file, indent=4, ensure_ascii=False)
-
-def append_to_csv(data, work_id):
-    if not os.path.exists(CSV_SAVE_PATH):
-        os.makedirs(CSV_SAVE_PATH)
-    csv_filename = os.path.join(CSV_SAVE_PATH, 'works.csv')
-    
-    file_exists = os.path.isfile(csv_filename)
-
-    if file_exists:
-        with open(csv_filename, 'r', newline='', encoding='utf-8') as csv_file:
-            reader = csv.reader(csv_file)
-            if any(row[0] == work_id for row in reader):
-                return
-
-    with open(csv_filename, 'a', newline='', encoding='utf-8') as csv_file:
-        writer = csv.writer(csv_file)
-        
-        if not file_exists:
-            writer.writerow(data.keys())
-        
-        writer.writerow([work_id, *data.values()])
-
-async def fetch_json(url: str):
-    async with httpx.AsyncClient() as client:
-        response = await client.get(url, timeout=20)
-        if response.status_code == 200:
-            return response.json()
-        else:
-            response.raise_for_status()
-
-def remove_latex_expressions(text):
-    latex_pattern = r'\$'
-    cleaned_text = re.sub(latex_pattern, '', text)
-
-    return cleaned_text
-
-async def get_abstract_text(abstract_inverted_index: dict) -> str:
-    word_index = []
-    
-    for word, indices in abstract_inverted_index.items():
-        for index in indices:
-            word_index.append((word, index))
-    
-    word_index = sorted(word_index, key=lambda x: x[1])
-    abstract = ' '.join([word for word, _ in word_index])
-    cleaned_abstract = remove_latex_expressions(abstract)
-    
-    return cleaned_abstract
-
-@app.get("/get_work_abstract/{id}")
-async def get_work_abstract(id: str):
-    id = id.upper() if id.lower().startswith("w") else f"W{id}"
-    full_id = f"{OPENALEX_API_URL}/works/{id}"
-
-    try:
-        data = await fetch_json(full_id)
-        
-        if 'abstract_inverted_index' in data:
-            abstract_inverted_index = data['abstract_inverted_index']
-            abstract_text = await get_abstract_text(abstract_inverted_index)
-            return {"abstract": abstract_text}
-        else:
-            return {"error": "Abstract not found for this work"}
-    
-    except httpx.HTTPStatusError as exc:
-        raise HTTPException(status_code=exc.response.status_code, detail="Error fetching data from OpenAlex")
+#@app.get("/get_work_abstract/{id}")
+#async def get_work_abstract(id: str):
+#    id = id.upper() if id.lower().startswith("w") else f"W{id}"
+#    full_id = f"{OPENALEX_API_URL}/works/{id}"
+#
+#    try:
+#        data = await fetch_json(full_id)
+#        
+#        if 'abstract_inverted_index' in data:
+#            abstract_inverted_index = data['abstract_inverted_index']
+#            abstract_text = await get_abstract_text(abstract_inverted_index)
+#            return {"abstract": abstract_text}
+#        else:
+#            return {"error": "Abstract not found for this work"}
+#    
+#    except httpx.HTTPStatusError as exc:
+#        raise HTTPException(status_code=exc.response.status_code, detail="Error fetching data from OpenAlex")
 
 @app.post("/get_works/")
 async def get_works(body: Annotated[WorkRequest, Body(
         example={"work_ids": ["W2117692326", "W2502170112", "W0987654321"]}
     )]):
+    if len(body.work_ids) > 50:
+        raise HTTPException(
+            status_code=400,
+            detail=f"Maximum number of IDs exceeded. No more than 50 allowed."
+        )
     results = []
     for id in body.work_ids:
         id = id.upper() if id.lower().startswith("w") else f"W{id}"
@@ -190,7 +133,14 @@ async def get_works(body: Annotated[WorkRequest, Body(
                 "title": data.get("display_name")
             }
             
+            if 'abstract_inverted_index' in data:
+                abstract_inverted_index = data['abstract_inverted_index']
+                abstract_text = await get_abstract_text(abstract_inverted_index)
+            else:
+                abstract_text = "Abstract not found for this work"
             
+            work_data["abstract"] = abstract_text
+
             save_to_json(work_data, id)
             append_to_csv(work_data, id)
             
@@ -201,14 +151,11 @@ async def get_works(body: Annotated[WorkRequest, Body(
         except Exception as e:
             results.append({"error": f"Internal server error: {str(e)}"})
 
-
     return results
-
 
 @app.get("/")
 async def read_root():
     return {"message": "Welcome to the parser System API"}
-
 
 if __name__ == "__main__":
     port = 8081
